requirements.txt
pillow<=9.2.0
numpy<=1.21.0
streamlit==1.11.0
torch<=1.8.2+cu111
opencv-python-headless<=4.6.0.66
PyYAML<=6.0
# FROM YOLOV5 -------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torchvision>=0.8.1
tqdm>=4.64.0
protobuf<4.21.3  
# Logging -------------------------------------
tensorboard>=2.4.1
# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0
# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
